added more granularity to the multiplication so i can do more problem sizes to figure out the cross over

+++ src/2mm-hand.cpp
@@ -86,7 +86,7 @@ int main(int argc, char ** argv) {
   
   camp::idx_t n = 1000;
   if(argc > 1) {
-    n *= std::atoi(argv[1]);
+    n *= std::atof(argv[1]);
   }
   hand(n);
   hand(n);
